Catch2 testing setup
