+++ test/traits.hh
@@ -1,0 +1,19 @@
+#include <dune/common/fvector.hh>
+
+/**
+ * @brief Types for coordinates and range values
+ */
+template<typename DF, typename RF, unsigned int dimension>
+class GridTraits
+{
+public:
+  enum
+  {
+    dim = dimension
+  };
+
+  using RangeField = RF;
+  using Scalar = Dune::FieldVector<RF, 1>;
+  using DomainField = DF;
+  using Domain = Dune::FieldVector<DF, dim>;
+};
